Base point for finishing up state management refactor into Module.
