+++ src/app/auth/auth.module.ts
@@ -1,13 +1,14 @@
 // Module Dependencies
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
+import { ReactiveFormsModule } from "@angular/forms";
+
 import {AuthRoutingModule} from "./auth-routing.module";
 
 // Custom Components
-import { LoginFormComponent } from "./login-form/login-form.component";
+import { LoginFormComponent } from "./components/login-form/login-form.component";
 
 import { AuthGuardService } from "../core/auth-guard.service"
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 
 
 export const COMPONENTS = [
