+++ src/app/core/core.module.ts
@@ -1,3 +1,5 @@
+// Common Module for comment UI components, services, etc...
+
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
