+++ src/app/core/services/services.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { ServicesService } from './services.service';
+
+describe('ServicesService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [ServicesService]
+    });
+  });
+
+  it('should be created', inject([ServicesService], (service: ServicesService) => {
+    expect(service).toBeTruthy();
+  }));
+});
