+++ src/app/core/services/services.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class ServicesService {
+
+  constructor() { }
+
+}
