+++ src/app/routes.ts
@@ -7,10 +7,6 @@ import {PartViewComponent} from "./part-view/part-view.component";
 import {Routes} from '@angular/router';
 import {PartMachineComponent} from "./part-machine/part-machine.component";
 import {PartListViewComponent} from "./part-list-view/part-list-view.component";
-import {LoginPageComponent} from "./login-page/login-page.component";
-// import { AuthGuardService } from './core/auth-guard.service';
-import {LoginFormComponent} from "./auth/login-form/login-form.component";
-
 
 
 export const appRoutes: Routes = [
